api/pyproject.toml
[project]
name = "demo-app"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.2,<0.35.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "aiomysql (>=0.2.0,<0.3.0)",
    "peewee (>=3.17.9,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "tomlkit (>=0.13.2,<0.14.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.26.0"
aiosqlite = "^0.21.0"
httpx = "^0.28.1"

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"